archived_files/pdf_to_csv_converter/requirements.txt
streamlit
pandas
camelot-py[cv]
tabula-py
PyPDF2
openpyxl
pdf2image
pytesseract
pdfplumber
zipfile36 